Rename docs_aide module to docs
