+++ rust/agama-server/src/web/docs.rs
